Reformat source to comply with ES6 modules and produce bundles with rollupjs
- This changes the sources to ES6 modules format, all located in a sub
  directory src/. To build production bundles with UMD wrappers for
  compatibility with browsers, etc. we use rollup.js.
