+++ example/index.html
@@ -5,7 +5,7 @@
     <title>IsoBands example</title>
 
     <script src="https://d3js.org/d3.v3.min.js"></script>
-    <script src="../marchingsquares-isobands.js"></script>
+    <script src="../dist/marchingsquares.js"></script>
 
 </head>
 <body>
