+++ example/vienna.html
@@ -5,8 +5,8 @@
     <title>IsoBands example</title>
 
     <script src="https://d3js.org/d3.v3.min.js"></script>
-    <script src="../marchingsquares-isobands.js"></script>
-    <script src="../marchingsquares-isocontours.js"></script>
+    <script src="../dist/marchingsquares-isobands.js"></script>
+    <script src="../dist/marchingsquares-isocontours.js"></script>
 
     <style>
 div.tooltip {
