+++ src/main.js
@@ -1,0 +1,15 @@
+/*!
+* @license GNU Affero General Public License.
+* Copyright (c) 2015-2018 Ronny Lorenz <ronny@tbi.univie.ac.at>
+* v. 1.2.3
+* https://github.com/RaumZeit/MarchingSquares.js
+*/
+
+import isoLines from './isolines.js';
+import isoBands from './isobands.js';
+
+export {
+  isoLines as isoLines,
+  isoBands as isoBands,
+  isoLines as isoContours
+}
