+++ src/main_isobands.js
@@ -1,0 +1,12 @@
+/*!
+* @license GNU Affero General Public License.
+* Copyright (c) 2015-2018 Ronny Lorenz <ronny@tbi.univie.ac.at>
+* v. 1.2.3
+* https://github.com/RaumZeit/MarchingSquares.js
+*/
+
+import isoBands from './isobands.js';
+
+export {
+  isoBands as isoBands,
+}
